+++ pirri/GlobalVars.go
