+++ pirri/pirriDripNode.go
