+++ pirri/pirriStationSchedule.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"time"
 
-	"../data"
-	"../logging"
-	"../settings"
+	"github.com/vacovsky/pirrigo/data"
+	"github.com/vacovsky/pirrigo/logging"
+	"github.com/vacovsky/pirrigo/settings"
 	"go.uber.org/zap"
 )
