+++ static/pirri_logo.xcf
